Create project
This is a simple chat microservice written in Python that allows users to log in or register for the chat service. When the program is run, the user is presented with two options: "Log in" or "Register". If the user chooses to log in, they must enter their username and password, which are then verified against the information stored in the MySQL database. If the user exists in the database, they are greeted with a welcome message. If the information is incorrect, they receive an error message. If the user chooses to register, they must enter a new username and password, which are then added to the MySQL database if the username is not already taken. The app uses the mysql-connector-python library to interact with the MySQL database.